+++ src/index.ts
@@ -66,11 +66,11 @@ export function ofAction < T extends Action > (...allowedTypes: any[]) {
 let store;
 export const ngrxSelect = (s => store = s);
 
-export function Select(path: string): any {
+export function Select(path?: string): any {
   return function (target: any, name: string, descriptor: TypedPropertyDescriptor <any>): void {
     if (delete target[name]) {
       Object.defineProperty(target, name, {
-        get: () => store.select(state => getValue(state, path)),
+        get: () => store.select(state => getValue(state, path || name)),
         enumerable: true,
         configurable: true
       });