+++ src/module.ts
@@ -1,0 +1,7 @@
+import { NgModule } from "@angular/core";
+import { NgrxSelect } from "./select";
+
+@NgModule({
+  providers: [NgrxSelect]
+})
+export class NgrxActionsModule {}