+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# Master
+# 2.0.0 - 1/5/17
+- Break: Add module for proper DI of selects
 - Feature: Implied select name from property name
 
 # 1.2.0 - 1/3/17